♻️ Moved `build_merge_query()` to `utils.py`
